Much better 3d item conversion and folder managment
With that update, i think this can maybe convert 90% of any RP Items Models. ( i'm not sure, but with what i have tested with a small  number of items. ) 

The Uv Conversion can be better i think but it's impossible for now for any converter to know how to found easily Texture size if it's not specified...